Swagger + Octet/Stream 문제 상황에 대해 더 명확한 이해를 가지고 이를 녹여낼 수 있는 구조로 코드 리팩토링
