+++ src/main/java/com/stream/msgconverter/OctetStreamReadMsgConverter.java
@@ -1,0 +1,33 @@
+package com.stream.msgconverter;
+
+import java.lang.reflect.Type;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.http.converter.json.AbstractJackson2HttpMessageConverter;
+import org.springframework.stereotype.Component;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+@Component
+public class OctetStreamReadMsgConverter extends AbstractJackson2HttpMessageConverter {
+	@Autowired
+	public OctetStreamReadMsgConverter(ObjectMapper objectMapper) {
+		super(objectMapper, MediaType.APPLICATION_OCTET_STREAM);
+	}
+
+	@Override
+	public boolean canWrite(Class<?> clazz, MediaType mediaType) {
+		return false;
+	}
+
+	@Override
+	public boolean canWrite(Type type, Class<?> clazz, MediaType mediaType) {
+		return false;
+	}
+
+	@Override
+	protected boolean canWrite(MediaType mediaType) {
+		return false;
+	}
+}
